+++ devdoctor/__init__.py
@@ -3,3 +3,7 @@
 from __future__ import annotations
 
 __version__ = "1.2.0rc1"
+__author__ = "AxiomNode"
+__copyright__ = "Copyright (c) 2026 AxiomNode"
+__license__ = "MIT"
+__homepage__ = "https://axiomnode.tech/"